CupertinoTextSelectionToolbar widget split out with related stuff
