+++ packages/flutter/test/cupertino/text_selection_toolbar_test.dart
@@ -1,0 +1,4 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
